Switch SemanticRoleLabeler metric to SrlEvalScorer.
